Refactor src/components/EventListCard/EventListCard.module.css
