+++ src/components/EventListCard/EventListCard.tsx
@@ -1,6 +1,6 @@
 import React, { useState } from 'react';
 import { useTranslation } from 'react-i18next';
-import styles from './EventListCard.module.css';
+import styles from '../../style/app.module.css';
 import { Navigate, useParams } from 'react-router-dom';
 import EventListCardModals from './EventListCardModals';
 import type { InterfaceRecurrenceRule } from 'utils/recurrenceUtils';
@@ -74,14 +74,14 @@ function eventListCard(props: InterfaceEventListCardProps): JSX.Element {
   return (
     <>
       <div
-        className={styles.cards}
+        className={styles.cardsEventListCard}
         style={{
           backgroundColor: '#d9d9d9',
         }}
         onClick={showViewModal}
         data-testid="card"
       >
-        <div className={styles.dispflex}>
+        <div className={styles.dispflexEventListCard}>
           <h2 className={styles.eventtitle}>
             {props.eventName ? <>{props.eventName}</> : <>Dogs Care</>}
           </h2>
